adding type definitions
