+++ package.json
@@ -26,7 +26,8 @@
   "exports": {
     ".": {
       "require": "./dist/index.js",
-      "import": "./dist/esm/index.js"
+      "import": "./dist/esm/index.js",
+      "types": "./dist/index.d.ts"
     },
     "./package.json": "./package.json"
   },
